Moved karyotype to anoph
